+++ core/src/dat/datrpg/entities/City.java
@@ -5,10 +5,10 @@ import java.util.ArrayList;
 
 public class City implements Serializable {
     public int mapRadius;
-    public int[][][] mapArray;
+    public byte[][][] mapArray;
     public ArrayList<House> houses;
 
-    public City(int mapRadius, int[][][] mapArray, ArrayList<House> houses) {
+    public City(int mapRadius, byte[][][] mapArray, ArrayList<House> houses) {
         this.mapRadius = mapRadius;
         this.mapArray = mapArray;
         this.houses = houses;
